code is refactored into separate csv writer/reader interface + implementation, unit test is updated
