+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/helper/csv/CsvWriter.java
@@ -1,7 +1,9 @@
-package com.apollocurrency.aplwallet.apl.core.shard.helper;
+/*
+ * Copyright © 2018-2019 Apollo Foundation
+ */
 
-import java.io.IOException;
-import java.io.Reader;
+package com.apollocurrency.aplwallet.apl.core.shard.helper.csv;
+
 import java.io.Writer;
 import java.sql.Connection;
 import java.sql.ResultSet;
@@ -9,7 +11,13 @@ import java.sql.SQLException;
 
 import com.apollocurrency.aplwallet.apl.core.db.derived.MinMaxDbId;
 
-public interface CsvManager {
+/**
+ * Class is used for writing into CSV (comma separated values) file.
+ * Folder for stored and parsed CSV files should be specified in implementation class.
+ *
+ * @author yuriy.larin
+ */
+public interface CsvWriter {
 
     /**
      * Writes the result set to a file in the CSV format.
@@ -42,6 +50,11 @@ public interface CsvManager {
     int append(String outputFileName, ResultSet rs, MinMaxDbId minMaxDbId) throws SQLException;
 
     /**
+     * Method is used with 'append' mode for explicit releasing resources
+     */
+    void close();
+
+    /**
      * Writes the result set of a query to a file in the CSV format.
      *
      * @param conn the connection
@@ -53,36 +66,6 @@ public interface CsvManager {
      * @return the number of rows written
      */
     int write(Connection conn, String outputFileName, String sql, String charset, MinMaxDbId minMaxDbId) throws SQLException;
-
-    /**
-     * Reads from the CSV file and returns a result set. The rows in the result
-     * set are created on demand, that means the file is kept open until all
-     * rows are read or the result set is closed.
-     * <br />
-     * If the columns are read from the CSV file, then the following rules are
-     * used: columns names that start with a letter or '_', and only
-     * contain letters, '_', and digits, are considered case insensitive
-     * and are converted to uppercase. Other column names are considered
-     * case sensitive (that means they need to be quoted when accessed).
-     *
-     * @param inputFileName the file name, not NULL
-     * @param colNames or null if the column names should be read from the CSV file
-     * @param charset the charset or null to use the system default charset (see system property file.encoding)
-     * @return the result set
-     */
-    ResultSet read(String inputFileName, String[] colNames, String charset) throws SQLException;
-
-    /**
-     * Reads CSV data from a reader and returns a result set. The rows in the
-     * result set are created on demand, that means the reader is kept open
-     * until all rows are read or the result set is closed.
-     *
-     * @param reader the reader
-     * @param colNames or null if the column names should be read from the CSV
-     *            file
-     * @return the result set
-     */
-    ResultSet read(Reader reader, String[] colNames) throws IOException;
 
     /**
      * Set String for Option values as 'key='value'. Possible parameters with default values are following:
