+++ apl-core/src/main/java/com/apollocurrency/aplwallet/apl/core/shard/helper/jdbc/SimpleRowSource.java
@@ -1,3 +1,7 @@
+/*
+ * Copyright © 2018-2019 Apollo Foundation
+ */
+
 package com.apollocurrency.aplwallet.apl.core.shard.helper.jdbc;
 
 import java.sql.SQLException;
